로그인 버그 수정, redis 토큰 방식 수정

+++ src/main/kotlin/com/waffletoy/team1server/user/UserTokenUtil.kt
@@ -26,7 +26,7 @@ object UserTokenUtil {
 
     private fun generateAccessToken(userEntity: UserEntity): String {
         val now = Instant.now()
-        val accessExpiryDate = Date.from(now.plusSeconds(ACCESS_TOKEN_EXPIRATION_TIME.toLong()))
+        val accessExpiryDate = Date.from(now.plusSeconds(ACCESS_TOKEN_EXPIRATION_TIME))
 
         return Jwts.builder()
             .signWith(SECRET_KEY)
